Refactored much of the code, made the default implementation thread-safe
Replaced the internal collections of the default Lambus implementation
with thread-safe alternatives; a result of this is that subscription
operations now take time proportional to the number of links subscribed
to the bus.
