+++ src/test/java/EventTester.java
@@ -1,6 +1,6 @@
 import me.foundry.lambus.Lambus;
 import me.foundry.lambus.Link;
-import me.foundry.lambus.internal.SynchronousLambus;
+import me.foundry.lambus.internal.LambusImpl;
 import me.foundry.lambus.priority.Prioritized;
 import me.foundry.lambus.priority.Priority;
 
@@ -12,7 +12,7 @@ import java.lang.management.ThreadMXBean;
  */
 public class EventTester {
     private final ThreadMXBean bean = ManagementFactory.getThreadMXBean();
-    private final Lambus bus = new SynchronousLambus();
+    private final Lambus bus = new LambusImpl();
 
     public void testDispatchTime() {
         Link<BasicEvent> link = bus.subscribeDirect((BasicEvent e) -> {});
@@ -50,7 +50,7 @@ public class EventTester {
 
     public void testSubscriptionTime() {
         long last = bean.getCurrentThreadCpuTime();
-        for (int i = 0; i < 1000000; i++) {
+        for (int i = 0; i < 100000; i++) {
             bus.subscribeAll(this);
         }
         final long current = bean.getCurrentThreadCpuTime() - last;
@@ -64,6 +64,10 @@ public class EventTester {
         bus.post(new BasicEvent());
         bus.unsubscribeAll(this);
         bus.post(new BasicEvent());
+    }
+
+    public void testConstruction() {
+        bus.subscribeDirect((BasicEvent e) -> {});
     }
 
 }
