+++ src/test/java/Main.java
@@ -6,8 +6,9 @@ public class Main {
     public static void main(String[] args) {
         final EventTester et = new EventTester();
         //et.testDispatchTime();
-        et.testPriorityHandling();
+        //et.testPriorityHandling();
         //et.testSubscriptionTime();
-        //et.testUnsubscription();
+        et.testUnsubscription();
+        //et.testConstruction();
     }
 }
